test(app-core): consolidate duplicate API and CLI test suite pairs (#28792)
