+++ packages/app-core/src/cli/profile-utils.test.ts
@@ -20,6 +20,8 @@ describe("profile-utils", () => {
       expect(isValidProfileName("_leading-underscore")).toBe(false);
       expect(isValidProfileName("invalid/slash")).toBe(false);
       expect(isValidProfileName("with space")).toBe(false);
+      expect(isValidProfileName("a;rm")).toBe(false);
+      expect(isValidProfileName("a".repeat(64))).toBe(true);
       expect(isValidProfileName("a".repeat(65))).toBe(false);
     });
   });
